refactored helpers
