+++ test/helpers/bytes-helpers.ts
@@ -1,5 +1,8 @@
 import { BigNumberish } from "ethers";
-import { addHexPrefix } from "./block-helpers";
+
+export function addHexPrefix(str: string): string {
+  return `0x${str}`;
+}
 
 export function reverseBytes(str: string) {
   if (str.slice(0, 2) == "0x") str = str.slice(2);
@@ -12,18 +15,6 @@ export function reverseByte(byte: string): string {
   const padded = binary.padStart(8, "0");
 
   return padded.split("").reverse().join("");
-}
-
-export function parseCuint(data: string, offset: number): [bigint, number] {
-  if (data.slice(offset, offset + 2) == "0x") data = data.slice(offset + 2);
-
-  const firstByte = parseInt(data.slice(offset, offset + 2), 16);
-
-  if (firstByte < 0xfd) return [BigInt(reverseBytes(data.slice(offset, offset + 2))), 2];
-  if (firstByte == 0xfd) return [BigInt(reverseBytes(data.slice(offset + 2, offset + 6))), 6];
-  if (firstByte == 0xfe) return [BigInt(reverseBytes(data.slice(offset + 2, offset + 10))), 10];
-
-  return [BigInt(reverseBytes(data.slice(offset + 2, offset + 18))), 18];
 }
 
 export function reverseUint32(decimalNumber: BigNumberish): BigInt {
